IN-63 refactor: wrap the entire application in next-redux-wrapper
