+++ src/app/providers/store/index.ts
@@ -1,3 +1,2 @@
-export { StoreProvider } from './StoreProvider'
 export { useAppDispatch, useAppSelector, useAppStore } from './store'
-/* export type { AppDispatch, AppStore, RootState } from './store' */
+export { type AppStore, type RootState, wrapper } from './store'
